Move websockets from legacy to asyncio
